Use `attr_name` instead of `_name`
The `nameko.extensions.DependencyProvider.bind` method sets the
`attr_name` attribute of the dependency provider to the attribute name
of the service class that it is bound to. This is more reliable than
passing the name as a parameter to the dependency provider (or setting
it via a metaclass on the service class).